requirements.txt
blis
catalogue
certifi
charset-normalizer
click
community
confection
contourpy
cycler
cymem
Flask
fonttools
fuzzywuzzy
idna
itsdangerous
Jinja2
joblib
kiwisolver
langcodes
MarkupSafe
matplotlib
murmurhash
networkx==2.8.8
numpy
packaging
pandas
pathy
Pillow
pydantic
pyparsing
PyTDC
python-dateutil
pytz
rdkit-pypi
requests
scikit-learn
scipy
seaborn
six
sklearn
smart-open
srsly
thinc
threadpoolctl
torch==1.13.1
tqdm
typer
typing_extensions
urllib3
wasabi
Werkzeug
mlflow
